its compiling again on the raspi

+++ src/AttitudeDeterminator/ComplimentaryFilter.cpp
@@ -3,7 +3,7 @@
 using namespace boost;
 using namespace boost::posix_time;
 
-ComplimentaryFilter::ComplimentaryFilter(float t_K_gyro){
+ComplimentaryFilter::ComplimentaryFilter(double t_K_gyro){
     if(K_gyro > 1 || K_gyro < 0){
         BOOST_LOG_TRIVIAL(info) << "WARNING: invalid value for K_gyro";
     }
@@ -13,12 +13,12 @@ ComplimentaryFilter::ComplimentaryFilter(float t_K_gyro){
     curTime.reset();
     pitch = 0;
 
-    accBuffer.reset(new circular_buffer<float>(ACC_BUFFER_SIZE));
-    accBuffer.reset(new circular_buffer<float>(GYRO_BUFFER_SIZE));
+    accBuffer.reset(new circular_buffer<double>(ACC_BUFFER_SIZE));
+    accBuffer.reset(new circular_buffer<double>(GYRO_BUFFER_SIZE));
 }
 
-float ComplimentaryFilter::filter(float accReading, float gyroReading){
-    float gyro, acc;
+double ComplimentaryFilter::filter(double accReading, double gyroReading){
+    double gyro, acc;
     bufferValues(accReading, gyroReading, &acc, &gyro);
 
     if(prevTime.get() == 0){ //TODO i dont want to check this every time.
@@ -29,8 +29,8 @@ float ComplimentaryFilter::filter(float accReading, float gyroReading){
     //Integrate gyro reading:
     time_duration diff = *curTime - *prevTime;
     unsigned long delta_t = diff.total_milliseconds();
-    float dt = static_cast<float>(delta_t)/1000;
-    float G = gyro*dt + pitch;
+    double dt = static_cast<double>(delta_t)/1000;
+    double G = gyro*dt + pitch;
 
     prevTime.swap(curTime); //save time for next cycle
     pitch = K_gyro * G + K_acc * acc; //save pitch
@@ -38,15 +38,15 @@ float ComplimentaryFilter::filter(float accReading, float gyroReading){
     return pitch;
 }
 
-void ComplimentaryFilter::bufferValues(float accReading, float gyroReading, float* acc, float* gyro){
+void ComplimentaryFilter::bufferValues(double accReading, double gyroReading, double* acc, double* gyro){
     accBuffer.get()->push_back(accReading);
     gyroBuffer.get()->push_back(gyroReading);
     *acc = calcAverage(*accBuffer.get());
     *gyro = calcAverage(*gyroBuffer.get());
 }
 
-float ComplimentaryFilter::calcAverage(circular_buffer<float> buffer) {
-    float total = 0;
+double ComplimentaryFilter::calcAverage(circular_buffer<double> buffer) {
+    double total = 0;
     for(int i = buffer.size() - 1; i >= 0; --i) {
         total += buffer[i];
     }
